03-12-2014

+++ WpfCopyApplication/Model/DataReplacement.cs
@@ -10,8 +10,10 @@ namespace WpfCopyApplication.Model
     {
         public string Path { get; set; }
         public DateTime Date { get; set; }
-        public int Size { get; set; }
+        public long Size { get; set; }
         public string Hash { get; set; }
         public string PathTargetDirectory { get; set; }
+
+
     }
 }
